fix: Se actualiza la vista y se agrega gestor de correo administrativo

+++ database/migrations/2021_03_12_003307_create_usuarios_table.php
@@ -17,7 +17,7 @@ class CreateUsuariosTable extends Migration
             $table->id();
             $table->string('nombres');
             $table->string('apellidos');
-            $table->integer('cedula')->unique();
+            $table->bigInteger('cedula')->unique();
             $table->string('email')->unique();
             $table->string('pais');
             $table->string('direccion');
